Split plugin to separate files
