+++ src/compiler/transforms/index.js
@@ -1,1 +1,3 @@
 export * from "./pre";
+export * from "./post";
+export * from "./jsxElement";
