+++ src/compiler/transforms/pre.js
@@ -7,10 +7,12 @@ import { shared } from "../shared";
 const createPre = (options) => {
   /**
    * @this {import('@babel/core').PluginPass}
-   * @param {import('@babel/core').BabelFile} path
+   * @param {import('@babel/core').BabelFile} file
    */
-  function pre(path) {
+  function pre(file) {
     shared.set("inlineRuntime", false);
+    shared.set("enableCommentOptions", false);
+    shared.set("enableStringMode", false);
     shared.set("importSource", "grim-jsx/dist/runtime.js");
 
     const { inuse, babel } = shared();
@@ -23,6 +25,14 @@ const createPre = (options) => {
     if (isObject(options)) {
       if ("importSource" in options && options.importSource) {
         shared.set("importSource", options.importSource);
+      }
+
+      if ("enableCommentOptions" in options && options.enableCommentOptions) {
+        shared.set("enableCommentOptions", options.enableCommentOptions);
+      }
+
+      if ("enableStringMode" in options && options.enableStringMode) {
+        shared.set("enableStringMode", options.enableStringMode);
       }
 
       if ("inlineRuntime" in options && options.inlineRuntime) {
